requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
numpy==1.24.3
pandas==2.0.3
python-dotenv==0.19.0
supabase==1.0.3
gunicorn==20.1.0
eptr2==1.0.1
